osa4 23/23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,32 +1,48 @@
 const blogsRouter = require('express').Router()
-const { request } = require('express')
 const Blog = require('../models/blog')
-const { response } = require('../app')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
+const user = require('../models/user')
 
-blogsRouter.get('/', (request, response) => {
-    Blog
+blogsRouter.get('/', async (request, response) => {
+    const blogs = await Blog
       .find({})
-      .then(blogs => {
-        response.json(blogs)
-      })
+      .populate('user', { username: 1, name: 1 })
+    response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
+
   const body = request.body
+  const user = request.user
   
   const blog = await new Blog({
     author: body.author,
     title: body.title,
     url: body.url,
-    likes: body.likes
+    likes: body.likes,
+    user: user._id
   })
   const savedBlog = await blog.save()
+
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+  if ( blog.user.toString() === user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'unauthorized user' })
+  }
+
+  
 })
 
 blogsRouter.put('/:id', async (request, response) => {
